TD has an additional trick
For files of years 2020, 2021 and 2022, there is another link that has to be used.

The links in href attribute of JSON file download incomplete files.

These special links are in the description attribute of JSON.

And there is one more trick.
There is a link in the description attribute for the year of 2012.
This links is broken!
So if the year is 2012 the link in the description must be ignored and the href link must be used.